requirements.txt
anyio==4.5.2
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.0
nbclient==0.10.1
nbconvert==7.16.5
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prompt-toolkit==3.0.48
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.20.2
